Unify API function signatures and pass in Msg
This makes the `Model` always the first argument to our API functions,
and then also forces us to pass in our errorMsg and msg to each function
call.

On the surface, this makes it more verbose - however, we really ought
not have the API itself explicitly choosing what functions are emitted.
It just seems to be bad form to me.